Moved lod_run to the ot2_client package and some reorganization
